Implement cargo clippy suggestions

+++ src/main.rs
@@ -108,7 +108,7 @@ fn main() {
     cpu.load(game_code);
     cpu.reset();
 
-    let mut screen_state = [0 as u8; 32 * 3 * 32];
+    let mut screen_state = [0_u8; 32 * 3 * 32];
     let mut rng = rand::thread_rng();
 
     // run the game cycle
